chore(shared): checkpoint, squashthis
